@@ -1,12 +1,11 @@
 name: beacons_plugin
 description: This plugin makes it easier to scan nearby iBeacons and range them in flutter app. Scaned beacons will provide proximity values.
-version: 1.0.12
-author: Umair Adil
+version: 2.0.0
 homepage: https://github.com/umair13adil/simple_beacons_flutter
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ^1.10.0
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.10.0"
 
 dependencies:
   flutter:
@@ -14,7 +13,7 @@ dependencies:
 
 dev_dependencies:
   flutter_test:
-    sdk: flutter
+    sdk: flutter 
 
 flutter:
   plugin:
