@@ -1,9 +1,13 @@
-import 'package:flutter/material.dart';
 import 'dart:async';
 import 'dart:io' show Platform;
-import 'package:beacons_plugin/beacons_plugin.dart';
 
-void main() => runApp(MyApp());
+import 'package:beacons_plugin/beacons_plugin.dart';
+import 'package:flutter/material.dart';
+
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  runApp(MyApp());
+}
 
 class MyApp extends StatefulWidget {
   @override
@@ -32,6 +36,16 @@ class _MyAppState extends State<MyApp> {
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
+    if (Platform.isAndroid) {
+      //Prominent disclosure
+      await BeaconsPlugin.setDisclosureDialogMessage(
+          title: "Need Location Permission",
+          message: "This app collects location data to work with beacons.");
+
+      //Only in case, you want the dialog to be shown again. By Default, dialog will never be shown if permissions are granted.
+      //await BeaconsPlugin.clearDisclosureDialogShowFlag(false);
+    }
+
     BeaconsPlugin.listenToBeacons(beaconEventsController);
 
     await BeaconsPlugin.addRegion(
@@ -60,14 +74,14 @@ class _MyAppState extends State<MyApp> {
     if (Platform.isAndroid) {
       BeaconsPlugin.channel.setMethodCallHandler((call) async {
         if (call.method == 'scannerReady') {
-          await BeaconsPlugin.startMonitoring;
+          await BeaconsPlugin.startMonitoring();
           setState(() {
             isRunning = true;
           });
         }
       });
     } else if (Platform.isIOS) {
-      await BeaconsPlugin.startMonitoring;
+      await BeaconsPlugin.startMonitoring();
       setState(() {
         isRunning = true;
       });
@@ -96,38 +110,26 @@ class _MyAppState extends State<MyApp> {
               SizedBox(
                 height: 20.0,
               ),
-              Visibility(
-                visible: isRunning,
-                child: RaisedButton(
-                  onPressed: () async {
-                    if (Platform.isAndroid) {
-                      await BeaconsPlugin.stopMonitoring;
-
-                      setState(() {
-                        isRunning = false;
-                      });
-                    }
-                  },
-                  child: Text('Stop Scanning', style: TextStyle(fontSize: 20)),
-                ),
+              ElevatedButton(
+                onPressed: () async {
+                  if(isRunning)
+                  {
+                    await BeaconsPlugin.stopMonitoring();
+                  }
+                  else
+                  {
+                    initPlatformState();
+                    await BeaconsPlugin.startMonitoring();
+                  }
+                  setState(() {
+                    isRunning = !isRunning;
+                  });
+                },
+                child: Text(isRunning?'Stop Scanning':'Start Scanning', style: TextStyle(fontSize: 20)),
               ),
               SizedBox(
                 height: 20.0,
               ),
-              Visibility(
-                visible: !isRunning,
-                child: RaisedButton(
-                  onPressed: () async {
-                    initPlatformState();
-                    await BeaconsPlugin.startMonitoring;
-
-                    setState(() {
-                      isRunning = true;
-                    });
-                  },
-                  child: Text('Start Scanning', style: TextStyle(fontSize: 20)),
-                ),
-              )
             ],
           ),
         ),
