@@ -1,13 +1,21 @@
 package com.umair.beacons_plugin_example
 
-import com.umair.beacons_plugin.BeaconsActivity
-import io.flutter.embedding.engine.FlutterEngine
-import io.flutter.plugins.GeneratedPluginRegistrant
+import com.umair.beacons_plugin.BeaconsPlugin
+import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity : BeaconsActivity(){
+class MainActivity : FlutterActivity(){
 
-    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
-        super.configureFlutterEngine(flutterEngine)
-        GeneratedPluginRegistrant.registerWith(flutterEngine)
+    override fun onPause() {
+        super.onPause()
+
+        //Start Background service to scan BLE devices
+        BeaconsPlugin.startBackgroundService(this)
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        //Stop Background service, app is in foreground
+        BeaconsPlugin.stopBackgroundService(this)
     }
 }
